+++ trakt/calendar.py
@@ -1,11 +1,13 @@
 """Interfaces to all of the Calendar objects offered by the Trakt.tv API"""
-from .tv import TVEpisode
 from .core import get
+from .movies import Movie
+from .tv import TVEpisode
 from .utils import now, airs_date
 
 __author__ = 'Jon Nappi'
-__all__ = ['Calendar', 'PremiereCalendar', 'ShowCalendar', 'SeasonCalendar',
-           'MovieCalendar']
+__all__ = ['Calendar', 'PremiereCalendar', 'MyPremiereCalendar',
+           'ShowCalendar', 'MyShowCalendar', 'SeasonCalendar',
+           'MySeasonCalendar', 'MovieCalendar', 'MyMovieCalendar']
 
 
 class Calendar(object):
@@ -23,21 +25,25 @@ class Calendar(object):
             days
         """
         super(Calendar, self).__init__()
-        self.date, self.days, self._episodes = date or now(), days, []
+        self.date, self.days, self._calendar = date or now(), days, []
         self._get()
+
+    def __getitem__(self, key):
+        """Pass index requests through to the internal _calendar object"""
+        return self._calendar.__getitem__(key)
 
     def __iter__(self):
         """Custom iterator for iterating over the episodes in this Calendar"""
-        return iter(self._episodes)
+        return iter(self._calendar)
 
     def __len__(self):
         """Returns the length of the episodes list in this calendar"""
-        return len(self._episodes)
+        return len(self._calendar)
 
     def __str__(self):
         """str representation of this Calendar"""
         from pprint import pformat
-        return pformat(self._episodes)
+        return pformat(self._calendar)
     __repr__ = __str__
 
     @property
@@ -52,48 +58,91 @@ class Calendar(object):
 
     def _build(self, data):
         """Build the calendar"""
-        self._episodes = []
-        for date in data:
-            episodes = data.get(date, [])
-            for episode in episodes:
-                show = episode.get('show', {}).get('title')
-                season = episode.get('episode', {}).get('season')
-                ep = episode.get('episode', {}).get('number')
-                data = {'airs_at': airs_date(episode.get('airs_at')),
-                        'episode_ids': episode.get('episode').get('ids'),
-                        'title': episode.get('episode', {}).get('title')}
-                self._episodes.append(TVEpisode(show, ep, season,
-                                                episode_data=data))
-        self._episodes = sorted(self._episodes, key=lambda x: x.airs_at)
+        self._calendar = []
+        for episode in data:
+            show = episode.get('show', {}).get('title')
+            season = episode.get('episode', {}).get('season')
+            ep = episode.get('episode', {}).get('number')
+            e_data = {'airs_at': airs_date(episode.get('first_aired')),
+                      'episode_ids': episode.get('episode').get('ids'),
+                      'title': episode.get('episode', {}).get('title')}
+            self._calendar.append(TVEpisode(show, ep, season, **e_data))
+        self._calendar = sorted(self._calendar, key=lambda x: x.airs_at)
 
 
 class PremiereCalendar(Calendar):
     """All shows premiering during the time period specified."""
 
     def __init__(self, date=None, days=7):
-        self.url = 'calendars/shows/new'
+        self.url = 'calendars/all/shows/new'
         super(PremiereCalendar, self).__init__(date=date, days=days)
+
+
+class MyPremiereCalendar(Calendar):
+    """Personalized calendar of all shows premiering during the time period
+    specified.
+    """
+
+    def __init__(self, date=None, days=7):
+        self.url = 'calendars/my/shows/new'
+        super(MyPremiereCalendar, self).__init__(date=date, days=days)
 
 
 class ShowCalendar(Calendar):
     """TraktTV ShowCalendar"""
 
     def __init__(self, date=None, days=7):
-        self.url = 'calendars/shows'
+        self.url = 'calendars/all/shows'
         super(ShowCalendar, self).__init__(date=date, days=days)
+
+
+class MyShowCalendar(Calendar):
+    """Personalized TraktTV ShowCalendar"""
+
+    def __init__(self, date=None, days=7):
+        self.url = 'calendars/my/shows'
+        super(MyShowCalendar, self).__init__(date=date, days=days)
 
 
 class SeasonCalendar(Calendar):
     """TraktTV TV Show Season Premiere"""
 
     def __init__(self, date=None, days=7):
-        self.url = 'calendars/shows/premieres'
+        self.url = 'calendars/all/shows/premieres'
         super(SeasonCalendar, self).__init__(date=date, days=days)
 
 
-class MovieCalendar(Calendar):
-    """TraktTV Movie Calendar"""
+class MySeasonCalendar(Calendar):
+    """Personalized TraktTV TV Show Season Premiere"""
 
     def __init__(self, date=None, days=7):
-        self.url = 'calendars/movies'
+        self.url = 'calendars/my/shows/premieres'
+        super(MySeasonCalendar, self).__init__(date=date, days=days)
+
+
+class MovieCalendar(Calendar):
+    """TraktTV Movie Calendar. Returns all movies with a release date during
+    the time period specified.
+    """
+
+    def __init__(self, date=None, days=7):
+        self.url = 'calendars/all/movies'
         super(MovieCalendar, self).__init__(date=date, days=days)
+
+    def _build(self, data):
+        """Build the calendar of Movies"""
+        self._calendar = []
+        for movie in data:
+            m_data = movie.get('movie', {})
+            released = movie.get('released', None)
+            self._calendar.append(Movie(released=released, **m_data))
+
+        self._calendar = sorted(self._calendar, key=lambda x: x.released)
+
+
+class MyMovieCalendar(MovieCalendar):
+    """Personalized TraktTV Movie Calendar."""
+
+    def __init__(self, date=None, days=7):
+        self.url = 'calendars/my/movies'
+        super(MyMovieCalendar, self).__init__(date=date, days=days)